# Docs improvements: Build-time Syntax highlighting, scroll perf improvements, docs dev/deployment fixes