--- a/package.json
+++ b/package.json
@@ -63,6 +63,8 @@
     "postcss-nested": "^3.0.0",
     "postcss-remtopx": "^1.0.0",
     "postcss-svg": "^2.1.0",
+    "prism": "^4.1.2",
+    "prismjs": "^1.15.0",
     "readable-stream": "^2.3.5",
     "run-sequence": "^1.2.2",
     "svgmin": "^0.1.0",