--- a/tasks/dev.js
+++ b/tasks/dev.js
@@ -45,11 +45,11 @@ function watch() {
     'docs/**/*.yml',
     'topdoc/lib/template.pug',
     'topdoc/lib/index.js',
-    '../topdoc/resources/js/*.js'
+    'topdoc/resources/js/*.js'
   ], gulp.series('reload-docs'));
 
   gulp.watch([
-    '../topdoc/resources/css/*.css'
+    'topdoc/resources/css/*.css'
   ], gulp.series('reload-docs-css'));
 
   gulp.watch('icons/*.svg', gulp.series('reload-icons'));