## Description

* Does build-time syntax highlighting with Prism instead of at run-time with Rainbow
* Removes the `pushState` on scroll
* Removes some dead code
* Removes an unused normalize.css
* Fixes livereload when working on docs resources
* Fixes gh-pages generation

This is not a complete solution to the problems described in the issues below. This change offers only marginal improvements to load time and scroll speed, but they took me a handful of minutes, so we've got that going for us.

## Related Issue

#7 - Docs load slow

#6 - Slow scrolling in docs

#26 - Remove `pushState` on scroll

#37 - Livereload broken for docs resources

#39 - gh-pages redirects missing

## Motivation and Context

This speeds up the loading of the page, and makes scrolling a little bit smoother. It also fixes a couple issues with docs generation that were introduced recently.

## How Has This Been Tested?

A bit of inspection in dev tools reveals marginal improvements in scroll jerkiness, and time to initial display has been improved by around 15% (6 seconds to 5 seconds), with time to complete load (including syntax highlighted code samples) improved by 50% (from 10 seconds to 5 seconds).

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- Docs fix (doesn't affect distributed package)