@lazd 2018-12-05
Discussed with @benjamind and @GarthDB and this is ready to merge.